requirements.txt
boto3
pandas
psycopg2
sqlalchemy
python-dotenv